.github/ci/requirements.txt
requests==2.32.3
google-genai>=1.4.0